feat: Item Freshness
Bringing in more control about how fresh (in hours) an item needs to be
to get be considered "fresh" and worth to be tooted.

Thus for news bots this value can be one way to avoid having older news
tootet at a later point in time.

Furthermore untooted feed items are deleted from the DB if they are
already stale.