feat(mastodon): Adding Content-Warnings
With this we are adding a new feature to the bot that is configurable
through the settings.json file.

It is a regex matching of expressions within the RSS text to check
whether it needs to be labeled by a content warning (short CW).

The order in which the cws are written in the settings matter. The first
CW matched within the text will win and will be added to the toot.

If none are matching, the toot is send without a cw.